requirements.txt
pandas
pillow
requests
tqdm
torch
torchvision
sentence-transformers
faiss-cpu
fastapi
uvicorn